Sync: PSR-4 the sync queue, buffer and utils
